requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-telegram-bot==20.3
apscheduler==3.10.1
requests==2.28.2
python-dotenv==1.0.0
